+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 CHANGELOG
 =========
 
-1.10.0
+unreleased
 ----------
+- Use generic error with console log when a payment method fails to set up
+- Fix issue where Mastercard was displayed as MasterCard
+
+1.10.0
+------
 - Enable Venmo support
 - Enable Google Pay support
 - Fix issue where non utf-8 encoded sites would show strange characters for card placeholders